第一版本的sync_wait完成
