+++ .gitignore
@@ -2,6 +2,6 @@
 out
 build
 .vscode
-.vscode/
+
