+++ CMakeLists.txt
@@ -2,6 +2,7 @@ cmake_minimum_required(VERSION 3.28.0)
 project(coro)
 set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CMAKE_BUILD_TYPE Debug)
 
 enable_testing()
